Refactor test function in _test with single options parameter including reporter, output coverage information, and 'applyContextTestCases' for text cases applied from environment; Call test function from task test_cover_jira_integration; Refactor location of jiraIssueManager to prevent gulpLoadTasks parsing file;
