+++ lib/utils/jiraIssueManager.js
