+++ tasks/_test.js
@@ -10,7 +10,6 @@
  * @param {bunyan} context.logger - A logger matching the bunyan API
  */
 module.exports = function testTasks(gulp, context) {
-  const jiraIssueManager = require("./lib/jiraIssueManager");
   const vasync = require("vasync");
   var mocha = require("gulp-mocha");
   var mkdirp = require("mkdirp");
@@ -78,7 +77,14 @@ module.exports = function testTasks(gulp, context) {
   };
 
 
-  var test = function test(reporter, outputCoverageReports) {
+  var test = function test(options = {
+    "reporter": "spec",
+    "outputCoverageReports": false,
+    "applyContextTestCases": true
+  }) {
+    var reporter = options.reporter;
+    var outputCoverageReports = options.outputCoverageReports;
+    var applyContextTestCases = options.applyContextTestCases;
     var cwd = context.cwd;
     var pkg = context.package;
     var directories = pkg.directories;
@@ -101,7 +107,10 @@ module.exports = function testTasks(gulp, context) {
       }
     }, glob.sync(sourceGlobStr));
 
-
+    if (applyContextTestCases && context.argv.length === 2) {
+      process.env.YADDA_FEATURE_GLOB = context.argv[1];
+      logger.info("Set process.env.YADDA_FEATURE_GLOB=" + process.env.YADDA_FEATURE_GLOB);
+    }
 
     if (outputCoverageReports) {
       return gulp.src(path.resolve(process.cwd(), directories.test + "/test.js"), {"read": false})
@@ -188,9 +197,6 @@ module.exports = function testTasks(gulp, context) {
    */
   gulp.task("test_cover", gulp.series(
     "instrument",
-    function() {
-      return applyContextTestCases(null);
-    },
     function testCoverTask() {
       var cwd = context.cwd;
       var pkg = context.package;
@@ -202,9 +208,17 @@ module.exports = function testTasks(gulp, context) {
       //make sure the Reports directory exists - required for mocha-bamboo-reporter-bgo
       mkdirp.sync(path.join(cwd, directories.reports));
       if (process.env.CI) {
-        return test("spec", true);
+        return test({
+          "reporter": "spec",
+          "outputCoverageReports": true,
+          "applyContextTestCases": true
+        });
       }
-      return test("mocha-bamboo-reporter-bgo", true);
+      return test({
+        "reporter": "mocha-bamboo-reporter-bgo",
+        "outputCoverageReports": true,
+        "applyContextTestCases": true
+      });
     }
   ));
 
@@ -216,26 +230,29 @@ module.exports = function testTasks(gulp, context) {
    * @member {Gulp} test_cover
    * @return {through2} stream
    */
-  gulp.task("test_cover_no_cov_report", gulp.series(
-    function() {
-      return applyContextTestCases(null);
-    },
-    function testCoverNoCovReportTask() {
-      var cwd = context.cwd;
-      var pkg = context.package;
-      var directories = pkg.directories;
-      var MOCHA_FILE_NAME = 'unit-mocha-tests' + (process.env.SELENIUM_PORT ? "-" + process.env.SELENIUM_PORT : "");
+  gulp.task("test_cover_no_cov_report", function testCoverNoCovReportTask() {
+    var cwd = context.cwd;
+    var pkg = context.package;
+    var directories = pkg.directories;
+    var MOCHA_FILE_NAME = 'unit-mocha-tests' + (process.env.SELENIUM_PORT ? "-" + process.env.SELENIUM_PORT : "");
 
-      //results file path for mocha-bamboo-reporter-bgo
-      process.env.MOCHA_FILE = path.join(cwd, directories.reports, MOCHA_FILE_NAME + ".json");
-      //make sure the Reports directory exists - required for mocha-bamboo-reporter-bgo
-      mkdirp.sync(path.join(cwd, directories.reports));
-      if (process.env.CI) {
-        return test("spec", false);
-      }
-      return test("mocha-bamboo-reporter-bgo", false);
-    })
-  );
+    //results file path for mocha-bamboo-reporter-bgo
+    process.env.MOCHA_FILE = path.join(cwd, directories.reports, MOCHA_FILE_NAME + ".json");
+    //make sure the Reports directory exists - required for mocha-bamboo-reporter-bgo
+    mkdirp.sync(path.join(cwd, directories.reports));
+    if (process.env.CI) {
+      return test({
+        "reporter": "spec",
+        "outputCoverageReports": false,
+        "applyContextTestCases": true
+      });
+    }
+    return test({
+      "reporter": "mocha-bamboo-reporter-bgo",
+      "outputCoverageReports": false,
+      "applyContextTestCases": true
+    });
+  });
 
   /**
    * A gulp build task to run test steps and calculate test coverage (but not output test coverage to prevent
@@ -404,40 +421,28 @@ module.exports = function testTasks(gulp, context) {
    * @member {Gulp} test
    * @return {through2} stream
    */
-  gulp.task("test", gulp.series(
-    function() {
-      return applyContextTestCases(null);
-    },
-    function testTask() {
-      return test("spec", true);
-    })
-  );
+  gulp.task("test", function testTask() {
+    return test({
+      "reporter": "spec",
+      "outputCoverageReports": true,
+      "applyContextTestCases": true
+    });
+  });
 
-  const applyContextTestCases = (jiraTestCases) => {
-
-    if (R.isEmpty(jiraTestCases) || R.isNil(jiraTestCases)) {
-      logger.info('No Jira issue key has been provided. Applying context parameters.');
-
-      if (context.argv.length === 2) {
-        process.env.YADDA_FEATURE_GLOB = context.argv[1];
-        logger.info("Set process.env.YADDA_FEATURE_GLOB=" + process.env.YADDA_FEATURE_GLOB);
-      }
-    }
-  };
-  
   /**
-   * A gulp build task to determine test cases to run. 
-   * First it will try to find a JIRA issue key in a branch name. 
+   * A gulp build task to determine test cases to run.
+   * First it will try to find a JIRA issue key in a branch name.
    * If found, it will search for its components and use them as test cases identifiers.
    * Otherwise, it will search for provided parameters in the context.
    * Test steps results will be output using spec reporter.
    * @member {Gulp} test_cover_jira_integration
    * @return {through2} stream
-   */  
+   */
   gulp.task("test_cover_jira_integration", () => {
+    const jiraIssueManager = require("../lib/utils/jiraIssueManager");
     vasync.waterfall([
       function(callback) {
-        jiraIssueManager.getJiraOauthClient(callback);      
+        jiraIssueManager.getJiraOauthClient(callback);
       },
       function(jiraOauthClient, callback) {
         jiraIssueManager.getJiraIssue(jiraOauthClient, callback);
@@ -447,12 +452,11 @@ module.exports = function testTasks(gulp, context) {
       },
       function getTestCasesToRun(jiraTestCases, callback) {
         process.env.YADDA_FEATURE_GLOB = jiraTestCases;
-        applyContextTestCases(jiraTestCases);
         callback(null, "Sucessfully finished test cases selection");
       }
-  
+
     ], function (error, result) {
-  
+
       if (!R.isEmpty(error) && !R.isNil(error)) {
         throw new Error(error);
       }
@@ -460,6 +464,18 @@ module.exports = function testTasks(gulp, context) {
       if (!R.isEmpty(result) && !R.isNil(result)) {
         logger.info(result);
       }
+      if (process.env.CI) {
+        return test({
+          "reporter": "spec",
+          "outputCoverageReports": false,
+          "applyContextTestCases": true
+        });
+      }
+      return test({
+        "reporter": "mocha-bamboo-reporter-bgo",
+        "outputCoverageReports": true,
+        "applyContextTestCases": false
+      });
     });
   });
 };
